api/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
grpcio>=1.34.0
grpcio-tools>=1.34.0
protobuf>=3.15.0